+++ cassandra/row_parser.pyx
@@ -13,13 +13,14 @@
 # limitations under the License.
 
 from cassandra.parsing cimport ParseDesc, ColumnParser
+from cassandra.policies import ColDesc
 from cassandra.obj_parser import TupleRowParser
 from cassandra.deserializers import make_deserializers
 
 include "ioutils.pyx"
 
 def make_recv_results_rows(ColumnParser colparser):
-    def recv_results_rows(self, f, int protocol_version, user_type_map, result_metadata):
+    def recv_results_rows(self, f, int protocol_version, user_type_map, result_metadata, column_encryption_policy):
         """
         Parse protocol data given as a BytesIO f into a set of columns (e.g. list of tuples)
         This is used as the recv_results_rows method of (Fast)ResultMessage
@@ -28,11 +29,12 @@ def make_recv_results_rows(ColumnParser colparser):
 
         column_metadata = self.column_metadata or result_metadata
 
-        self.column_names = [c[2] for c in column_metadata]
-        self.column_types = [c[3] for c in column_metadata]
+        self.column_names = [md[2] for md in column_metadata]
+        self.column_types = [md[3] for md in column_metadata]
 
-        desc = ParseDesc(self.column_names, self.column_types, make_deserializers(self.column_types),
-                         protocol_version)
+        desc = ParseDesc(self.column_names, self.column_types, column_encryption_policy,
+                        [ColDesc(md[0], md[1], md[2]) for md in column_metadata],
+                        make_deserializers(self.column_types), protocol_version)
         reader = BytesIOReader(f.read())
         try:
             self.parsed_rows = colparser.parse_rows(reader, desc)